src/main/java/id/ac/ui/cs/advprog/eshop/repository/PaymentRepository.java
package id.ac.ui.cs.advprog.eshop.repository;

import id.ac.ui.cs.advprog.eshop.model.Payment;

import java.util.ArrayList;
import java.util.List;

public class PaymentRepository {

    private List<Payment> paymentData = new ArrayList<>();
    public Payment save(Payment payment) {
       int i = 0;
       for (Payment savePayment : paymentData){
           if (savePayment.getId().equals(payment.getId())){
               paymentData.remove(i);
               paymentData.add(i, payment);
               return payment;
           }
           i += 1;
       }

       paymentData.add(payment);
       return payment;
    }

    public Payment findById (String id){
        for (Payment savePayment: paymentData){
            if (savePayment.getId().equals(id)){
                return  savePayment;
            }
        }
        return null;
    }

    public List<Payment> getAllPayment (){
        List<Payment> result = new ArrayList<>();
        for (Payment savedPayment : paymentData){
            {
                result.add(savedPayment);
            }
        }

        return  result;
    }
}
